# src/CreatorPantry.Tests/Api/Snapshots/openapi-v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "CreatorPantry API",
    "description": "Browser clients call these routes through CreatorPantry.Gateway with a session cookie. Errors are RFC 9457 problem details carrying a stable \u0060code\u0060 and a \u0060traceId\u0060.",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/workspaces/{workspaceSlug}/recipes/{recipeId}/ai-proposals/{aiProposalId}": {
      "get": {
        "tags": [
          "AiProposals"
        ],
        "summary": "Where a requested proposal has got to, and the proposal itself once there is one.",
        "description": "Polling before the proposal exists is a \u0060200\u0060 carrying a status, not a \u0060404\u0060 \u2014 the request\nresource exists from the moment it was accepted, and reporting it missing would make a normal poll look\nlike a failure.",
        "parameters": [
          {
            "name": "workspaceSlug",
            "in": "path",
            "description": "Where a requested proposal has got to, and the proposal itself once there is one.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recipeId",
            "in": "path",
            "description": "Where a requested proposal has got to, and the proposal itself once there is one.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "aiProposalId",
            "in": "path",
            "description": "The id the request returned. It is the operation\u0027s id: a request exists from the moment it is accepted,\nlong before a proposal does, so there would be nothing else to poll. The proposal carries its own\nidentity inside the reply for a disposition to act on.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AiProposalStatus"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProposalStatus"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProposalStatus"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "Problem details with a stable \u0060code\u0060 and \u0060traceId\u0060.",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bffSession": []
          }
        ]
      }
    },
    "/api/v1/workspaces/{workspaceSlug}/recipes/{recipeId}/ai-proposals/{aiProposalId}/disposition": {
      "post": {
        "tags": [
          "AiProposals"
        ],
        "summary": "Records what the creator decided about a proposal, and applies what they accepted.",
        "description": "\u0060200\u0060 rather than \u0060202\u0060: unlike requesting a proposal, this finishes before it returns. There is\nno provider call \u2014 the answer was generated long ago \u2014 and the recipe version, the per-change decisions,\nthe feedback and the operation\u0027s terminal status all commit in one transaction before the reply is written.\n\n\u0060Contributor\u0060 at the edge, because accepting changes writes a recipe version. A \u0060Viewer\u0060 cannot\nreject either, which is the deliberate cost of one policy per route: rejecting decides the fate of a\nworkspace\u0027s content, and the role that may not change a recipe should not be the role that closes a\nproposal against it.\n\nRetrying is safe and needs no \u0060Idempotency-Key\u0060. A proposal can be decided once; a repeat of the same\ndecision is recognised and answered without writing anything, and a repeat asking for a\ndifferent decision is a \u0060409\u0060 rather than a second version of the recipe.",
        "parameters": [
          {
            "name": "workspaceSlug",
            "in": "path",
            "description": "Records what the creator decided about a proposal, and applies what they accepted.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recipeId",
            "in": "path",
            "description": "Records what the creator decided about a proposal, and applies what they accepted.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "aiProposalId",
            "in": "path",
            "description": "The same id  takes \u2014 the request\u0027s. An operation has at most one proposal, so the request\nidentifies it without ambiguity, and one route segment meaning two different identities would be a trap.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "$ref": "#/components/parameters/AntiforgeryToken"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiProposalDisposition"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AiProposalDisposition"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/AiProposalDisposition"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AiProposalDisposition"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProposalDisposition"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProposalDisposition"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "Problem details with a stable \u0060code\u0060 and \u0060traceId\u0060.",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bffSession": []
          }
        ]
      }
    },
...
    "/api/v1/workspaces/{workspaceSlug}/proposal-explanation-requests/{requestId}": {
      "get": {
        "tags": [
          "ProposalExplanationRequests"
        ],
        "summary": "Where a requested explanation has got to, and the explanation itself once there is one.",
        "description": "Polling before the explanation exists is a \u0060200\u0060 carrying a status, not a \u0060404\u0060 -- the request\nresource exists from the moment it was accepted, and reporting it missing would make a normal poll look\nlike a failure.",
        "parameters": [
          {
            "name": "workspaceSlug",
            "in": "path",
            "description": "Where a requested explanation has got to, and the explanation itself once there is one.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "requestId",
            "in": "path",
            "description": "The id the request returned. It is the operation\u0027s id.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AiProposalStatus"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProposalStatus"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiProposalStatus"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "Problem details with a stable \u0060code\u0060 and \u0060traceId\u0060.",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bffSession": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AiChangeDisposition": {
        "enum": [
          "Pending",
          "Accepted",
          "Rejected"
        ],
        "type": "string"
      },
      "AiChangeKind": {
        "enum": [
          "Unspecified",
          "Set",
          "Add",
          "Remove",
          "Move"
        ],
        "type": "string"
      },
      "AiChangeTargetKind": {
        "enum": [
          "Unspecified",
          "Recipe",
          "Ingredient",
          "IngredientGroup",
          "InstructionStep",
          "InstructionGroup",
          "Equipment",
          "AssetLink",
          "Tag",
          "RecipeConcept",
          "IngredientSubstitution",
          "RecipeReviewFinding",
          "ProposalExplanationItem"
        ],
        "type": "string"
      },
...
      "AiFailureCategory": {
        "enum": [
          "Unspecified",
          "Validation",
          "Quota",
          "TemplateUnavailable",
          "Provider",
          "RateLimited",
          "Timeout",
          "OutputSchemaInvalid",
          "DomainInvalid",
          "SafetyBlocked",
          "Cancelled",
          "LeaseAbandoned",
          "AccountSuspended"
        ],
        "type": "string"
      },
      "AiOperationScope": {
        "enum": [
          "Unspecified",
          "WholeRecipe",
          "Ingredients",
          "Instructions",
          "Metadata",
          "Media",
          "NotApplicable",
          "Advisory"
        ],
        "type": "string"
      },
      "AiOperationStatus": {
        "enum": [
          "Requested",
          "Running",
          "Proposed",
          "Accepted",
          "PartiallyAccepted",
          "Rejected",
          "Failed",
          "Expired"
        ],
        "type": "string"
      },
      "AiProposalDetail": {
        "required": [
          "proposalId",
          "outputSchemaVersion",
          "promptTemplateId",
          "promptTemplateVersion",
          "promptTemplateBodyChecksum",
          "providerName",
          "modelName",
          "createdAt",
          "changes",
          "warnings"
        ],
        "type": "object",
        "properties": {
          "proposalId": {
            "type": "string",
            "format": "uuid"
          },
          "outputSchemaVersion": {
            "type": "string"
          },
          "promptTemplateId": {
            "type": "string"
          },
          "promptTemplateVersion": {
            "type": "string"
          },
          "promptTemplateBodyChecksum": {
            "type": "string"
          },
          "providerName": {
            "type": "string"
          },
          "modelName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "changes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AiProposedChange"
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AiProposalWarning"
            }
          }
        }
      },
      "AiProposalDisposition": {
        "required": [
          "aiProposalRequestId",
          "status",
          "acceptedChangeCount",
          "rejectedChangeCount",
          "recipeVersionNumber",
          "decidedAt"
        ],
        "type": "object",
        "properties": {
          "aiProposalRequestId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/AiOperationStatus"
          },
          "acceptedChangeCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "rejectedChangeCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "recipeVersionNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "decidedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AiProposalStatus": {
        "required": [
          "aiProposalRequestId",
          "status",
          "taskType",
          "scope",
          "sourceVersionId",
          "requestedAt",
          "statusChangedAt",
          "failureCategory",
          "proposal"
        ],
        "type": "object",
        "properties": {
          "aiProposalRequestId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/AiOperationStatus"
          },
          "taskType": {
            "$ref": "#/components/schemas/AiTaskType"
          },
          "scope": {
            "$ref": "#/components/schemas/AiOperationScope"
          },
          "sourceVersionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "requestedAt": {
            "type": "string",
            "format": "date-time"
          },
          "statusChangedAt": {
            "type": "string",
            "format": "date-time"
          },
          "failureCategory": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AiFailureCategory"
              }
            ]
          },
          "proposal": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AiProposalDetail"
              }
            ]
          }
        }
      },
      "AiProposalWarning": {
        "required": [
          "kind",
          "message",
          "changeId"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/AiWarningKind"
          },
          "message": {
            "type": "string"
          },
          "changeId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
      "AiProposedChange": {
        "required": [
          "changeId",
          "changeKind",
          "targetKind",
          "targetId",
          "fieldName",
          "beforeValue",
          "afterValue",
          "proposedPosition",
          "disposition"
        ],
        "type": "object",
        "properties": {
          "changeId": {
            "type": "string",
            "format": "uuid"
          },
          "changeKind": {
            "$ref": "#/components/schemas/AiChangeKind"
          },
          "targetKind": {
            "$ref": "#/components/schemas/AiChangeTargetKind"
          },
          "targetId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "fieldName": {
            "type": [
              "null",
              "string"
            ]
          },
          "beforeValue": {
            "type": [
              "null",
              "string"
            ]
          },
          "afterValue": {
            "type": [
              "null",
              "string"
            ]
          },
          "proposedPosition": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "disposition": {
            "$ref": "#/components/schemas/AiChangeDisposition"
          }
        }
      },
...
      "AiTaskType": {
        "enum": [
          "Unspecified",
          "Diagnostic",
          "RecipeConcepts",
          "RecipeFirstDraft",
          "RecipeRevision",
          "IngredientSubstitution",
          "RecipeAdaptation",
          "RecipeReview",
          "ProposalExplanation"
        ],
        "type": "string"
      },
      "AiWarningKind": {
        "enum": [
          "Unspecified",
          "Assumption",
          "CulinaryCaution",
          "NonScalableLanguage",
          "UnverifiedClaim",
          "SafetyCaution",
          "UnresolvedQuestion",
          "Rationale",
          "Limitation"
        ],
        "type": "string"
      },
...
      "ProblemDetails": {
        "required": [
          "code",
          "traceId"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "code": {
            "type": "string",
            "description": "Stable, machine-readable error code, e.g. \u0060auth.registration.invalid\u0060 or \u0060not_found\u0060."
          },
          "traceId": {
            "type": "string",
            "description": "Correlation id for support and tracing."
          }
        }
      },
...
      "ValidationProblemDetails": {
        "required": [
          "code",
          "traceId"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "code": {
            "type": "string",
            "description": "Stable, machine-readable error code, e.g. \u0060auth.registration.invalid\u0060 or \u0060not_found\u0060."
          },
          "traceId": {
            "type": "string",
            "description": "Correlation id for support and tracing."
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Field errors keyed by camelCase request field name."
          }
        }
      },
...
    },
    "parameters": {
...
      "AntiforgeryToken": {
        "name": "X-XSRF-TOKEN",
        "in": "header",
        "description": "Antiforgery request token from GET /bff/antiforgery (or the sign-in response). Required by the gateway on every unsafe request; a missing or stale token is rejected with 400 csrf.invalid.",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
...
    "securitySchemes": {
      "bffSession": {
        "type": "apiKey",
        "description": "HttpOnly session cookie issued by the gateway (BFF). Browsers never handle tokens. Provisional name, finalized with the BFF session.",
        "name": "__Host-creatorpantry-session",
        "in": "cookie"
      }
    }
  },
...
}